requirements.txt
Django==5.1.1
djangorestframework==3.15.2
stripe==10.10.0
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0
beautifulsoup4==4.12.3
selenium==4.23.1
pandas==2.2.2
